--- packages/ui/src/views/connectors/GithubConnector.ts
@@ -1,5 +1,0 @@
-import { getConnectorBuilder } from './ConnectorPayloadBuilder';
-import { GithubConnectorSpec } from './schemas';
-
-// Get the builder for GitHub connector
-export const githubBuilder = getConnectorBuilder<GithubConnectorSpec>('Github');
