+++ packages/ui/src/views/connectors/harness-connectors/aws-kms-connector.tsx
@@ -2,14 +2,7 @@ import { InputConfigType, InputType } from '@views/unified-pipeline-studio/compo
 
 import { IFormDefinition } from '@harnessio/forms'
 
-import { ConnectorPayloadConfig, ConnectorSpec } from '../ConnectorPayloadBuilder'
-import {
-  AWS_KMS_CONNECTOR_IDENTIFIER,
-  createConnectorPayloadConverter,
-  CredTypeValues,
-  DelegateTypes,
-  IInputConfigWithConfigInterface
-} from '../types'
+import { CredTypeValues, DelegateTypes, IInputConfigWithConfigInterface } from '../types'
 
 export const AWS_KMS_CONNECTOR_DESCRIPTION = 'AWS KMS Connector'
 
@@ -29,33 +22,33 @@ const inputs: IInputConfigWithConfigInterface[] = [
   },
   {
     inputType: InputType.text,
-    path: `spec.spec.accessKey`,
+    path: `accessKey`,
     label: 'AWS - Access Key',
     isVisible: values => values?.credential === CredTypeValues.ManualConfig
   },
   {
     inputType: InputType.text,
-    path: `spec.spec.secretKey`,
+    path: `secretKey`,
     label: 'AWS - Secret Key',
     isVisible: values => values?.credential === CredTypeValues.ManualConfig
   },
   {
     inputType: InputType.text,
-    path: `spec.spec.roleArn`,
+    path: `roleArn`,
     label: 'Role ARN',
-    isVisible: values => values?.credential === CredTypeValues.ManualConfig
+    isVisible: values => values?.credential === CredTypeValues.AssumeRoleSTS
   },
   {
     inputType: InputType.text,
-    path: `spec.spec.externalName`,
+    path: `externalName`,
     label: 'External Id',
-    isVisible: values => values?.credential === CredTypeValues.ManualConfig
+    isVisible: values => values?.credential === CredTypeValues.AssumeRoleSTS
   },
   {
-    inputType: InputType.text,
-    path: `spec.spec.assumeStsRoleDuration`,
+    inputType: InputType.number,
+    path: `assumeStsRoleDuration`,
     label: 'Assumed Role duration',
-    isVisible: values => values?.credential === CredTypeValues.ManualConfig
+    isVisible: values => values?.credential === CredTypeValues.AssumeRoleSTS
   },
   {
     inputType: InputType.text,
@@ -80,73 +73,6 @@ const inputs: IInputConfigWithConfigInterface[] = [
   }
 ]
 
-// Convert AWS KMS payload back to form data
-interface AwsKmsConnectorSpec extends ConnectorSpec {
-  name: string
-  credential: string
-  awsArn: string
-  region: string
-  default?: boolean
-  executeOnDelegate?: boolean
-  iamRole?: string
-  accessKey?: string
-  secretKey?: string
-  delegateSelectors?: string[]
-  roleArn?: string
-  externalName?: string
-  assumeStsRoleDuration?: string
-}
-
-const convertAwsKmsPayloadToFormData = (payload: ConnectorPayloadConfig<ConnectorSpec>): Record<string, any> => {
-  // Extract fields from spec
-  const spec = payload.spec as AwsKmsConnectorSpec
-
-  // Basic form data
-  const formData: Record<string, any> = {
-    name: payload.name,
-    description: payload.description,
-    projectIdentifier: payload.projectIdentifier,
-    orgIdentifier: payload.orgIdentifier,
-    identifier: payload.identifier,
-    tags: payload.tags,
-    credential: spec.credential,
-    awsArn: spec.awsArn,
-    region: spec.region,
-    default: spec.default
-  }
-
-  // Add OIDC-specific fields
-  if (spec.credential === DelegateTypes.DELEGATE_OIDC && spec.iamRole) {
-    formData.iamRole = spec.iamRole
-  }
-
-  // Add credential-specific fields
-  switch (spec.credential) {
-    case CredTypeValues.ManualConfig:
-      formData.accessKey = spec.accessKey
-      formData.secretKey = spec.secretKey
-      break
-    case CredTypeValues.AssumeIAMRole:
-      formData.delegateSelectors = spec.delegateSelectors
-      break
-    case CredTypeValues.AssumeRoleSTS:
-      formData.delegateSelectors = spec.delegateSelectors
-      formData.roleArn = spec.roleArn
-      formData.externalName = spec.externalName
-      formData.assumeStsRoleDuration = spec.assumeStsRoleDuration
-      break
-  }
-
-  return formData
-}
-
 export const awsKmsConnectorFormDefinition: IFormDefinition<InputConfigType> = {
   inputs
 }
-
-// Create and export the AWS KMS connector payload converter
-export const awsKmsConnectorPayloadConverter = createConnectorPayloadConverter(
-  AWS_KMS_CONNECTOR_IDENTIFIER,
-  convertAwsKmsPayloadToFormData,
-  awsKmsConnectorFormDefinition
-)
